requirements.txt
pynput>=1.7.6
Pillow>=10.0.0
pywin32>=306
pystray>=0.19.4
